refactor(type): 타입을 파일로 따로 분리
