+++ src/domain/Restaurants.ts
@@ -1,10 +1,4 @@
-type Restaurant = {
-  category: string;
-  name: string;
-  distance: string;
-  description: string;
-  link: string;
-};
+import { Restaurant } from '../type/common';
 
 class Restaurants {
   #restaurants: Restaurant[] = [];
