+++ src/type/common.ts
@@ -1,0 +1,9 @@
+type Restaurant = {
+  category: string;
+  name: string;
+  distance: string;
+  description: string;
+  link: string;
+};
+
+export { Restaurant };
